+++ test/mocha/helpers.js
@@ -1,19 +1,10 @@
-/*
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
 const brAccount = require('bedrock-account');
 const database = require('bedrock-mongodb');
-const {promisify} = require('util');
-
-exports.getActors = async mockData => {
-  const actors = {};
-  for(const [key, record] of Object.entries(mockData.accounts)) {
-    actors[key] = await brAccount.getCapabilities({id: record.account.id});
-  }
-  return actors;
-};
 
 exports.prepareDatabase = async mockData => {
   await exports.removeCollections();
@@ -24,7 +15,7 @@ exports.removeCollections = async (
   collectionNames = [
     'account',
   ]) => {
-  await promisify(database.openCollections)(collectionNames);
+  await database.openCollections(collectionNames);
   for(const collectionName of collectionNames) {
     await database.collections[collectionName].deleteMany({});
   }
@@ -38,7 +29,7 @@ async function _insertTestData(mockData) {
   for(const record of records) {
     try {
       await brAccount.insert(
-        {actor: null, account: record.account, meta: record.meta || {}});
+        {account: record.account, meta: record.meta || {}});
     } catch(e) {
       if(e.name === 'DuplicateError') {
         // duplicate error means test data is already loaded