+++ test/test.config.js
@@ -1,11 +1,10 @@
-/*
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+/*!
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
 const {config} = require('bedrock');
 const path = require('path');
-const {permissions, roles} = config.permission;
 
 config.mocha.tests.push(path.join(__dirname, 'mocha'));
 
@@ -21,16 +20,5 @@ config['https-agent'].rejectUnauthorized = false;
 
 // this needs to be true in order to set auth stuff
 config['account-http'].autoLoginNewAccounts = true;
-
-roles['account.registered'] = {
-  id: 'account.registered',
-  label: 'Account Test Role',
-  comment: 'Role for Test User',
-  sysPermission: [
-    permissions.ACCOUNT_ACCESS.id,
-    permissions.ACCOUNT_UPDATE.id,
-    permissions.ACCOUNT_INSERT.id
-  ]
-};
 
 config.express.useSession = true;